Dodanie produktów, zmiana id z long na int, poprawiona metoda toString(), dodany nowy konstruktor dla klasy Computer,, zmiana w klasie Order konstruktora ZonedDateTime

+++ src/Electronics.java
@@ -1,5 +1,6 @@
 public class Electronics extends Product{
-    public Electronics(Long id, String productName, double price, int quantityAvailable) {
+    public Electronics(int id, String productName, double price, int quantityAvailable) {
         super(id, productName, price, quantityAvailable);
     }
+
 }
